+++ Package.swift
@@ -15,8 +15,8 @@ let package = Package(
     ],
     targets: [
         .binaryTarget(name: "FireworkVideo",
-                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v0.3.0/FireworkVideo-v0.3.0.xcframework.zip",
-                      checksum: "d8eca178d4692107ab4c76c204e50b8dafdaa3778b9e937f12eb4f4f24760cee"),
+                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v0.4.0/FireworkVideo-v0.4.0.xcframework.zip",
+                      checksum: "cfd7894407e6f76598e7edcbd22694dcc79a5b2c3a3edd606e5e65ced29d52f8")
     ]
     
 )