+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-03-18
+
+### Added
+
+- Video player tool tip user onboarding aides.
+- Video player tool tip can now be disabled.
+- When using a channel source, the video feed can be initialized with a playlist encoded ID.
+
 ## [0.3.0] - 2021-03-05
 
 ### Added